services/ingestor/requirements.txt
httpx==0.27.0
PyYAML==6.0.1
qdrant-client==1.7.3
